Fix routes for vlan interfaces under debian/ubuntu
Whenever a file has a . in the filename (for example when adding routes for bond0.1332 interface, it would then be route-bond0.1332) the route file is skipped. By changing the dots in the iface name for underscores, it is properly executed and handled and routes are also added for vlan interfaces.